Interfaz y lógica de visionado de grupos propios y modificaciones generales a otros servlets. (#3)
* Esqueleto de la aplicación

* Modelo de datos de User añadido

* Modelo de datos de Login añadido

* Modelo de las Listas de Deseos añadido

* Añadidos constructores a los modelos existentes (correción)

* Añadido el modelo de Agrupaciones

* Modelo de datos de Grupos añadido

* Añadido el parámetro fecha a la clase Grupos

* Añadido el parámetro fecha a la clase Grupos (ahora si, no guardé en el anterior)

* añadida interfaz de DAO de Agrupaciones (posiblemente ampliable)

* añadida interfaz de DAO de User (posiblemente ampliable)

* añadida interfaz de DAO de Login (posiblemente ampliable)

* Correciones y algunos constructores añadidos

* Añadido el resto de interfaces

* Correcciones

* Correcciones

* Implementación (incompleta) y con fallos

* Errores corregidos

* Más fallos corregidos y metidos metodos para borrar elementos de las listas y grupos

* Borrado de la cache los settings de Eclipse

* Metido .gitignore

* Servlet e interfaz casi completadas. Falta añadir botón para meter a más gente en un grupo

* Añadido boton de añadir persona a un grupo, funciones logicas, y adaptado el servlet de sorteo básico al uso de las funciones

+++ src/es/upm/dit/isst/amigos/dao/AgrupacionesDAO.java
@@ -15,7 +15,7 @@ public interface AgrupacionesDAO {
 	
 	public List<Agrupaciones> getAgrupacionesByUser(String user);
 	
-	public List<Agrupaciones> getAgrupacionesByGrupo(int grupo);
+	public List<Agrupaciones> getAgrupacionesByGrupo(Long grupo);
 	
 	public List<Agrupaciones> getAgrupByUserAndGrupo(String user, int grupo);
 }
